if the config delete_merged_file is false, do not delete the merged main file: https://stackoverflow.com/q/61973608/559676

+++ R/render.R
@@ -134,7 +134,7 @@ render_book = function(
   } else {
     render_cur_session(files, main, config, output_format, clean, envir, ...)
   }
-  file.remove(main)
+  if (!xfun::isFALSE(delete_main)) file.remove(main)
   res
 }
